Fix a bug where empty sets would cause toDynamo to generate invalid data
Amazon does not allow puts with the attribute name listed but an empty data
argument.  All collection attributes must contain at least _one_ item.
So for empty collections I just omit that attribute.

This seems good for the developer also, if the collection is empty the
attribute won't be there and when we someday call fromDynamo we'll correctly
get an empty collection.